Update dependency and change the frontend codes

+++ aspnetapp/WINGS/ClientApp/src/components/home/OperationList.tsx
@@ -1,34 +1,26 @@
 import React, { useState } from 'react';
-import { makeStyles, createStyles } from '@material-ui/styles';
-import Table from '@material-ui/core/Table';
-import TableBody from '@material-ui/core/TableBody';
-import TableCell from '@material-ui/core/TableCell';
-import TableContainer from '@material-ui/core/TableContainer';
-import TableHead from '@material-ui/core/TableHead';
-import TableRow from '@material-ui/core/TableRow';
-import Paper from '@material-ui/core/Paper';
-import IconButton from '@material-ui/core/IconButton';
-import Tooltip from '@material-ui/core/Tooltip';
-import InputIcon from '@material-ui/icons/Input';
-import StopIcon from '@material-ui/icons/Stop';
-import UndoIcon from '@material-ui/icons/Undo';
+import Table from '@mui/material/Table';
+import TableBody from '@mui/material/TableBody';
+import TableCell from '@mui/material/TableCell';
+import TableContainer from '@mui/material/TableContainer';
+import TableHead from '@mui/material/TableHead';
+import TableRow from '@mui/material/TableRow';
+import Paper from '@mui/material/Paper';
+import IconButton from '@mui/material/IconButton';
+import Tooltip from '@mui/material/Tooltip';
+import InputIcon from '@mui/icons-material/Input';
+import StopIcon from '@mui/icons-material/Stop';
+import UndoIcon from '@mui/icons-material/Undo';
 import { useDispatch, useSelector } from 'react-redux';
 import { getOpid } from '../../redux/operations/selectors';
-import { leaveOperationAction } from '../../redux/operations/actions';
+import { joinOperationAction, leaveOperationAction } from '../../redux/operations/actions';
 import { joinOperation } from '../../redux/operations/operations';
 import { Operation } from '../../models';
 import { RootState } from '../../redux/store/RootState';
-import { cyan } from "@material-ui/core/colors";
+import { cyan } from "@mui/material/colors";
 import ConfirmationDialog from '../common/ConfirmationDialog';
 import { openErrorDialogAction } from '../../redux/ui/actions';
-
-const useStyles = makeStyles( 
-  createStyles({
-    isJoining: {
-      backgroundColor: cyan[800]
-    },
-  })
-);
+import { AppDispatch } from '../../redux/store/store';
 
 export interface OperationListProps {
   operations: Operation[],
@@ -36,9 +28,8 @@ export interface OperationListProps {
 }
 
 const OperationList = (props: OperationListProps) => {
-  const classes = useStyles();
   const selector = useSelector((state: RootState) => state);
-  const dispatch = useDispatch();
+  const dispatch = useDispatch<AppDispatch>();
   const opid = getOpid(selector);
 
   const [open, setOpen] = useState(false);
@@ -49,8 +40,15 @@ const OperationList = (props: OperationListProps) => {
     setStopOperation(operation);
   }
 
-  const handleOkClick = async () => {
-    const res = await fetch(`/api/operations/${stopOperation?.id}`, {
+  const joinOperationWrapper = (operation: Operation) => {
+    dispatch(joinOperation(operation)).catch((error) => {
+      console.error("Error failed to join Operation: ", error);
+    })
+  };
+
+  const handleOk = async () => {
+    if (stopOperation === null) return;
+    const res = await fetch(`/api/operations/${stopOperation.id}`, {
       method: 'DELETE'
     })
     if (res.status === 204) {
@@ -63,6 +61,12 @@ const OperationList = (props: OperationListProps) => {
       dispatch(openErrorDialogAction(message));
     }
   }
+
+  const handleOkClick = () => {
+    handleOk().catch((error) => {
+      console.error("Error failed to fetch operation:", error);
+    })
+  };
 
   const handleDialogClose = () => {
     setOpen(false);
@@ -82,21 +86,21 @@ const OperationList = (props: OperationListProps) => {
               <TableCell>Path Number</TableCell>
               <TableCell>Comment</TableCell>
               <TableCell>Component</TableCell>
-              <TableCell className="t-row-icon-cell"/>
-              <TableCell className="t-row-icon-cell"/>
-              <TableCell className="t-row-icon-cell"/>
+              <TableCell className="t-row-icon-cell" />
+              <TableCell className="t-row-icon-cell" />
+              <TableCell className="t-row-icon-cell" />
             </TableRow>
           </TableHead>
           <TableBody>
             {props.operations.length > 0 && (
               props.operations.map(operation => (
-                <TableRow key={operation.id} className={`${opid === operation.id && classes.isJoining}`}>
-                  <TableCell style={{width: "150px"}}>{operation.pathNumber}</TableCell>
+                <TableRow key={operation.id} sx={opid === operation.id ? { backgroundColor: cyan[800] } : {}}>
+                  <TableCell style={{ width: "150px" }}>{operation.pathNumber}</TableCell>
                   <TableCell>{operation.comment}</TableCell>
                   <TableCell>{operation.component.name}</TableCell>
                   <TableCell>
                     <Tooltip title="Join">
-                      <IconButton className="t-row-icon-cell" onClick={() => dispatch(joinOperation(operation))}>
+                      <IconButton className="t-row-icon-cell" onClick={() => joinOperationWrapper(operation)}>
                         <InputIcon />
                       </IconButton>
                     </Tooltip>
@@ -122,7 +126,7 @@ const OperationList = (props: OperationListProps) => {
         </Table>
       </TableContainer>
       <ConfirmationDialog
-        open={open} onOkClick={async () => await handleOkClick()}
+        open={open} onOkClick={() => handleOkClick()}
         labelOk="Stop" onClose={handleDialogClose}
       >
         <p>Are your sure to stop</p>
